requirements/base.txt
Django==4.2.7
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-modeltranslation
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.26
django-phonenumber-field==7.2.0
Pillow==9.5.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
